requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.0.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography>=2.3
cupshelpers==1.0
defer==1.0.6
distro-info==0.18
Django==2.1.2
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.1
gunicorn==19.7.1
httplib2==0.9.2
idna==2.6
isort==4.3.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.10
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.1.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.2
python-debian==0.1.32
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.0.0
wadllib==1.3.2
wrapt==1.10.11
xkit==0.0.0
zope.interface==4.3.2
